refacto singlePost
